Updated upload endpoint
